fix: 마이페이지 통합
