+++ src/pages/index.tsx
@@ -1,9 +1,8 @@
 export { default as RegisterPage } from "./Auth/Register";
 export { default as LoginPage } from "./Auth/Login";
 export { default as MainPage } from "./Main";
-export { default as MyPageUpdate } from "./MyPage/MyPageUpdate";
 export { default as GamePage } from "./Game";
 export { default as MyReposPage } from "./MyRepository";
 export { default as GameProblem } from "./Game/GameProblem";
 export { default as MyPageRead } from "./MyPage/MyPageRead";
-export { default as Wait } from "./Wait"
+export { default as Wait } from "./Wait";
